Create README.md
Hello everyone. I am an elderly man of 75 years trying to learn how to programming ESP32 and ARDUINO. I need it for an alarmcontroller for my cabin in the mountains, which I can control from anywhere in the world. It must contain a burglar alarm and that it must control two relays for heat control. It must also have 3 inputs for temperature measurement.